added the "administration" module
Lost this at some point, but i figured it would be good to add it back
in.

If you want to make it work, turn off Toolbar, in drupal package. If
not you will have two toolbars.

Next turn on the administration module, all 3 boxes.

Save, and run the update script.

If you like the regular one, just don't turn on the administration
module.